Rename runtime-types -> runtime.types
