+++ src/types/index.ts
@@ -1,3 +1,3 @@
-export * from './runtime-types';
+export * from './runtime.types';
 export * from './cql-code-service.interfaces';
 export * from './cql-patient.interfaces';
